backend/requirements.txt
annotated-types==0.7.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Pydantic==0.12.0
Flask-PyMongo==3.0.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
pymongo==4.11.3
pytest==8.3.5
python-dotenv==1.1.0
typing-inspection==0.4.0
typing_extensions==4.13.0
Werkzeug==3.1.3
